refactor sequence_similarity and add tests
